# backend/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DicDoc Service",
        "title": "DicDoc Service",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/doctors/{id}/workdays": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получить  расписание врача по его идентификатору в таблице TimeTable",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "doctors"
                ],
                "summary": "Получение расписания врача по его идентификатору",
                "operationId": "get-timetable-by-doctor-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор врача",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Расписание",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.GetTimetableResponseDto"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    }
                }
            }
        },
        "/api/v1/locations": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получить кабинеты по идентификатору врача в таблице TimeTable",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "locations"
                ],
                "summary": "Получение рабочих кабинетов по идентификатору врача",
                "operationId": "get-locations-by-doctor-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор врача",
                        "name": "doctor_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о кабинетах",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.GetOfficeResponseDto"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ErrorDto": {
            "description": "Информация об ошибке (DTO)",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Ошибка",
                    "type": "string"
                }
            }
        },
...
        "dto.GetOfficeResponseDto": {
            "description": "Информация о кабинете при получении",
            "type": "object",
            "properties": {
                "floor": {
                    "description": "Этаж",
                    "type": "integer"
                },
                "id": {
                    "description": "Идентификатор кабинета",
                    "type": "integer"
                },
                "number": {
                    "description": "Номер кабинета",
                    "type": "integer"
                }
            }
        },
...
        "dto.GetTimetableResponseDto": {
            "description": "Информация о расписании при получении",
            "type": "object",
            "properties": {
                "doctorId": {
                    "description": "Идентификатор врача",
                    "type": "integer"
                },
                "id": {
                    "description": "Идентификатор записи в таблице",
                    "type": "integer"
                },
                "officeId": {
                    "description": "Идентификатор кабинета",
                    "type": "integer"
                },
                "workDay": {
                    "description": "Номер рабочего дня",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Authorization check",
            "type": "apiKey",
            "name": "auth-x",
            "in": "header"
        }
    }
}